+++ core/src/cats/postgresql.cc
@@ -201,18 +201,50 @@ const char* BareosDbPostgresql::OpenDatabase(JobControlRecord* jcr)
     port = NULL;
   }
 
+  std::vector<const char*> keys;
+  std::vector<const char*> values;
+  if (db_address_) {
+    keys.emplace_back("host");
+    values.emplace_back(db_address_);
+  }
+  if (port) {
+    keys.emplace_back("port");
+    values.emplace_back(port);
+  }
+  if (db_name_) {
+    keys.emplace_back("dbname");
+    values.emplace_back(db_name_);
+  }
+  if (db_user_) {
+    keys.emplace_back("user");
+    values.emplace_back(db_user_);
+  }
+  if (db_password_) {
+    keys.emplace_back("password");
+    values.emplace_back(db_password_);
+  }
+  keys.emplace_back("sslmode");
+  values.emplace_back("disable");
+
+  keys.emplace_back(nullptr);
+  values.emplace_back(nullptr);
+  ASSERT(keys.size() == values.size());
   // If connection fails, try at 5 sec intervals for 30 seconds.
   for (int retry = 0; retry < 6; retry++) {
-    db_handle_ = PQsetdbLogin(db_address_,   /* default = localhost */
-                              port,          /* default port */
-                              NULL,          /* pg options */
-                              NULL,          /* tty, ignored */
-                              db_name_,      /* database name */
-                              db_user_,      /* login name */
-                              db_password_); /* password */
+    db_handle_ = PQconnectdbParams(keys.data(), values.data(), true);
 
-    // If no connect, try once more in case it is a timing problem
+    // If connecting does not succeed, try again in case it was a timing
+    // problem
     if (PQstatus(db_handle_) == CONNECTION_OK) { break; }
+
+    const char* err = PQerrorMessage(db_handle_);
+    if (!err) { err = "unknown reason"; }
+    Dmsg1(50, "Could not connect to db: Err=%s\n", err);
+    Mmsg2(errmsg,
+          T_("Unable to connect to PostgreSQL server. Database=%s User=%s\n"
+             "Possible causes: SQL server not running; password incorrect; "
+             "server requires ssl; max_connections exceeded.\n(%s)\n"),
+          db_name_, db_user_, err);
 
     // free memory if not successful
     PQfinish(db_handle_);
@@ -221,19 +253,11 @@ const char* BareosDbPostgresql::OpenDatabase(JobControlRecord* jcr)
     Bmicrosleep(5, 0);
   }
 
-  Dmsg0(50, "pg_real_connect %s\n",
-        PQstatus(db_handle_) == CONNECTION_OK ? "ok" : "failed");
+  Dmsg0(50, "pg_real_connect %s\n", db_handle_ ? "ok" : "failed");
   Dmsg3(50, "db_user=%s db_name=%s db_password=%s\n", db_user_, db_name_,
         (db_password_ == NULL) ? "(NULL)" : db_password_);
 
-  if (PQstatus(db_handle_) != CONNECTION_OK) {
-    Mmsg2(errmsg,
-          T_("Unable to connect to PostgreSQL server. Database=%s User=%s\n"
-             "Possible causes: SQL server not running; password incorrect; "
-             "max_connections exceeded.\n(%s)\n"),
-          db_name_, db_user_, PQerrorMessage(db_handle_));
-    return errmsg;
-  }
+  if (!db_handle_) { return errmsg; }
 
   connected_ = true;
   if (!CheckTablesVersion(jcr)) { return errmsg; }