+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - mtx-changer: make mandatory test mt-st versus cpio-mt [PR #2256]
 - packaging: set all `*.conf.examples` as %config(noreplace) [PR #2268]
 - disable bareosfd-python3-module-test on FreeBSD [PR #2278]
+- postgresql: require non-ssl connection [PR #2272]
 
 ### Documentation
 - docs: fix grpc-fd plugin call [PR #2068]
@@ -142,6 +143,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #2264]: https://github.com/bareos/bareos/pull/2264
 [PR #2268]: https://github.com/bareos/bareos/pull/2268
 [PR #2270]: https://github.com/bareos/bareos/pull/2270
+[PR #2272]: https://github.com/bareos/bareos/pull/2272
 [PR #2275]: https://github.com/bareos/bareos/pull/2275
 [PR #2278]: https://github.com/bareos/bareos/pull/2278
 [unreleased]: https://github.com/bareos/bareos/tree/master